Move ref.Val <-> api.expr.Value converters to io.go
This allows users of the go CEL library to consume this code. This
change also adds a round tripping test case for the converters

Move ref.value <-> api.expr.Value converters to io.go

Initial commit to test compilation

Fix compilation issues caused by previous commit

Add round tripping test for ref.Val <-> Value converter

Fix compilation issues and ensure that io_test runs and works
