+++ common/types/provider.go
@@ -22,8 +22,6 @@ import (
 	"github.com/google/cel-go/common/types/pb"
 	"github.com/google/cel-go/common/types/ref"
 	"github.com/google/cel-go/common/types/traits"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/reflect/protoreflect"
 
@@ -31,7 +29,6 @@ import (
 	anypb "google.golang.org/protobuf/types/known/anypb"
 	dpb "google.golang.org/protobuf/types/known/durationpb"
 	structpb "google.golang.org/protobuf/types/known/structpb"
-	rpcpb "google.golang.org/genproto/googleapis/rpc/status"
 	tpb "google.golang.org/protobuf/types/known/timestamppb"
 )
 
@@ -443,133 +440,6 @@ func nativeToValue(a ref.TypeAdapter, value interface{}) (ref.Val, bool) {
 		}
 	}
 	return nil, false
-}
-
-// RefValueToExprValue converts between ref.Val and exprpb.ExprValue.
-func RefValueToExprValue(res ref.Val, err error) (*exprpb.ExprValue, error) {
-	if err != nil {
-		s := status.Convert(err).Proto()
-		return &exprpb.ExprValue{
-			Kind: &exprpb.ExprValue_Error{
-				Error: &exprpb.ErrorSet{
-					Errors: []*rpcpb.Status{s},
-				},
-			},
-		}, nil
-	}
-	if IsUnknown(res) {
-		return &exprpb.ExprValue{
-			Kind: &exprpb.ExprValue_Unknown{
-				Unknown: &exprpb.UnknownSet{
-					Exprs: res.Value().([]int64),
-				},
-			}}, nil
-	}
-	v, err := RefValueToValue(res)
-	if err != nil {
-		return nil, err
-	}
-	return &exprpb.ExprValue{
-		Kind: &exprpb.ExprValue_Value{Value: v}}, nil
-}
-
-// RefValueToValue converts between ref.Val and Value.
-// The ref.Val must not be error or unknown.
-func RefValueToValue(res ref.Val) (*exprpb.Value, error) {
-	switch res.Type() {
-	case BoolType:
-		return &exprpb.Value{
-			Kind: &exprpb.Value_BoolValue{BoolValue: res.Value().(bool)}}, nil
-	case BytesType:
-		return &exprpb.Value{
-			Kind: &exprpb.Value_BytesValue{BytesValue: res.Value().([]byte)}}, nil
-	case DoubleType:
-		return &exprpb.Value{
-			Kind: &exprpb.Value_DoubleValue{DoubleValue: res.Value().(float64)}}, nil
-	case IntType:
-		return &exprpb.Value{
-			Kind: &exprpb.Value_Int64Value{Int64Value: res.Value().(int64)}}, nil
-	case ListType:
-		l := res.(traits.Lister)
-		sz := l.Size().(Int)
-		elts := make([]*exprpb.Value, 0, int64(sz))
-		for i := Int(0); i < sz; i++ {
-			v, err := RefValueToValue(l.Get(i))
-			if err != nil {
-				return nil, err
-			}
-			elts = append(elts, v)
-		}
-		return &exprpb.Value{
-			Kind: &exprpb.Value_ListValue{
-				ListValue: &exprpb.ListValue{Values: elts}}}, nil
-	case MapType:
-		mapper := res.(traits.Mapper)
-		sz := mapper.Size().(Int)
-		entries := make([]*exprpb.MapValue_Entry, 0, int64(sz))
-		for it := mapper.Iterator(); it.HasNext().(Bool); {
-			k := it.Next()
-			v := mapper.Get(k)
-			kv, err := RefValueToValue(k)
-			if err != nil {
-				return nil, err
-			}
-			vv, err := RefValueToValue(v)
-			if err != nil {
-				return nil, err
-			}
-			entries = append(entries, &exprpb.MapValue_Entry{Key: kv, Value: vv})
-		}
-		return &exprpb.Value{
-			Kind: &exprpb.Value_MapValue{
-				MapValue: &exprpb.MapValue{Entries: entries}}}, nil
-	case NullType:
-		return &exprpb.Value{
-			Kind: &exprpb.Value_NullValue{}}, nil
-	case StringType:
-		return &exprpb.Value{
-			Kind: &exprpb.Value_StringValue{StringValue: res.Value().(string)}}, nil
-	case TypeType:
-		typeName := res.(ref.Type).TypeName()
-		return &exprpb.Value{Kind: &exprpb.Value_TypeValue{TypeValue: typeName}}, nil
-	case UintType:
-		return &exprpb.Value{
-			Kind: &exprpb.Value_Uint64Value{Uint64Value: res.Value().(uint64)}}, nil
-	case DurationType:
-		d, ok := res.Value().(time.Duration)
-		if !ok {
-			return nil, status.New(codes.InvalidArgument, "Expected time.Duration").Err()
-		}
-		any, err := anypb.New(dpb.New(d))
-		if err != nil {
-			return nil, err
-		}
-		return &exprpb.Value{
-			Kind: &exprpb.Value_ObjectValue{ObjectValue: any}}, nil
-	case TimestampType:
-		t, ok := res.Value().(time.Time)
-		if !ok {
-			return nil, status.New(codes.InvalidArgument, "Expected time.Time").Err()
-		}
-		any, err := anypb.New(tpb.New(t))
-		if err != nil {
-			return nil, err
-		}
-		return &exprpb.Value{
-			Kind: &exprpb.Value_ObjectValue{ObjectValue: any}}, nil
-	default:
-		// Object type
-		pb, ok := res.Value().(proto.Message)
-		if !ok {
-			return nil, status.New(codes.InvalidArgument, "Expected proto message").Err()
-		}
-		any, err := anypb.New(pb)
-		if err != nil {
-			return nil, err
-		}
-		return &exprpb.Value{
-			Kind: &exprpb.Value_ObjectValue{ObjectValue: any}}, nil
-	}
 }
 
 func msgSetField(target protoreflect.Message, field *pb.FieldDescription, val ref.Val) error {
